CTCTRADERS-144:move new AddEventPage to events package
