+++ app/pages/events/AddEventPage.scala
@@ -14,32 +14,14 @@
  * limitations under the License.
  */
 
-package forms
+package pages.events
 
-import forms.behaviours.BooleanFieldBehaviours
-import play.api.data.FormError
+import pages.QuestionPage
+import play.api.libs.json.JsPath
 
-class AddEventFormProviderSpec extends BooleanFieldBehaviours {
+case object AddEventPage extends QuestionPage[Boolean] {
 
-  val requiredKey = "addEvent.error.required"
-  val invalidKey  = "error.boolean"
+  override def path: JsPath = JsPath \ toString
 
-  val form = new AddEventFormProvider()()
-
-  ".value" - {
-
-    val fieldName = "value"
-
-    behave like booleanField(
-      form,
-      fieldName,
-      invalidError = FormError(fieldName, invalidKey)
-    )
-
-    behave like mandatoryField(
-      form,
-      fieldName,
-      requiredError = FormError(fieldName, requiredKey)
-    )
-  }
+  override def toString: String = "addEvent"
 }
