+++ conf/views/events/addEvent.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.AddEventController.onSubmit(mrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.events.AddEventController.onSubmit(mrn, mode).url }}" method="post">
 
           {{ csrf() | safe }}
