+++ test/forms/events/EventCountryFormProviderSpec.scala
@@ -14,27 +14,34 @@
  * limitations under the License.
  */
 
-package forms
+package forms.events
 
-import forms.behaviours.BooleanFieldBehaviours
-import forms.events.IsTranshipmentFormProvider
+import forms.behaviours.StringFieldBehaviours
 import play.api.data.FormError
 
-class IsTranshipmentFormProviderSpec extends BooleanFieldBehaviours {
+class EventCountryFormProviderSpec extends StringFieldBehaviours {
 
-  val requiredKey = "isTranshipment.error.required"
-  val invalidKey  = "error.boolean"
+  val requiredKey = "eventCountry.error.required"
+  val lengthKey   = "eventCountry.error.length"
+  val maxLength   = 2
 
-  val form = new IsTranshipmentFormProvider()()
+  val form = new EventCountryFormProvider()()
 
   ".value" - {
 
     val fieldName = "value"
 
-    behave like booleanField(
+    behave like fieldThatBindsValidData(
       form,
       fieldName,
-      invalidError = FormError(fieldName, invalidKey)
+      stringsWithMaxLength(maxLength)
+    )
+
+    behave like fieldWithMaxLength(
+      form,
+      fieldName,
+      maxLength = maxLength,
+      lengthError = FormError(fieldName, lengthKey, Seq(maxLength))
     )
 
     behave like mandatoryField(
