+++ test/forms/events/EventReportedFormProviderSpec.scala
@@ -14,35 +14,26 @@
  * limitations under the License.
  */
 
-package forms
+package forms.events
 
-import forms.behaviours.StringFieldBehaviours
-import forms.events.EventPlaceFormProvider
+import forms.behaviours.BooleanFieldBehaviours
 import play.api.data.FormError
 
-class EventPlaceFormProviderSpec extends StringFieldBehaviours {
+class EventReportedFormProviderSpec extends BooleanFieldBehaviours {
 
-  val requiredKey = "eventPlace.error.required"
-  val lengthKey   = "eventPlace.error.length"
-  val maxLength   = 35
+  val requiredKey = "eventReported.error.required"
+  val invalidKey  = "error.boolean"
 
-  val form = new EventPlaceFormProvider()()
+  val form = new EventReportedFormProvider()()
 
   ".value" - {
 
     val fieldName = "value"
 
-    behave like fieldThatBindsValidData(
+    behave like booleanField(
       form,
       fieldName,
-      stringsWithMaxLength(maxLength)
-    )
-
-    behave like fieldWithMaxLength(
-      form,
-      fieldName,
-      maxLength = maxLength,
-      lengthError = FormError(fieldName, lengthKey, Seq(maxLength))
+      invalidError = FormError(fieldName, invalidKey)
     )
 
     behave like mandatoryField(
